passing src prop to Modal and displaying photo
useState("") to set initial value of SelectedPhotoSrc and pass new value as the already declared mediaURL variable in the rendered <Photo> component.

In Modal.jsx, put a new prop src as the value of the src attribute in the <img> tag. Then in App,jsx, pass the src prop with value as the initial value of selectedPhotoSrc variable.

The expected behavior is: the src prop in Modal is initially set as an empty value, displaying no photo. when the img is clicked, the handleClick function fire that opens the modal and src prop's value is updated to mediaURL which contains the API photo url.